livemart/requirements.txt
django
djangorestframework
django-filter
dj-rest-auth
django-allauth
Pillow
requests